Tests moved to class _NewItemTest, Delete old class
