+++ src/test/java/_NewItemTest.java
@@ -133,4 +133,16 @@ public class _NewItemTest extends BaseTest {
             Assert.assertTrue(icon.isEnabled());
         }
     }
+
+    @Test
+    public void testCheckBreadcrumbs() {
+        getDriver().findElement(By.xpath("//a[@href='/view/all/newJob']")).click();
+        String title = getDriver().getTitle();
+        List<WebElement> breadcrumbs = getDriver().findElements(By.xpath("//ul[@id='breadcrumbs']/li"));
+
+        Assert.assertEquals(title, "New Item [Jenkins]");
+        Assert.assertEquals(breadcrumbs.get(0).getText(), "Dashboard");
+        Assert.assertEquals(breadcrumbs.get(2).getText(), "All");
+
+    }
 }
